fixing dtype bugs

+++ llava/model/multimodal_encoder/clip_encoder.py
@@ -48,11 +48,15 @@ class CLIPVisionTower(nn.Module):
             image_features = []
             for image in images:
                 image_forward_out = self.vision_tower(image.to(device=self.device, dtype=self.dtype).unsqueeze(0), output_hidden_states=True)
-                image_feature = self.feature_select(image_forward_out).to(image.dtype)
+                # Keep vision tower dtype instead of converting back to input dtype
+                # This is critical for DP training where input dtype might differ from model dtype
+                image_feature = self.feature_select(image_forward_out).to(self.dtype)
                 image_features.append(image_feature)
         else:
             image_forward_outs = self.vision_tower(images.to(device=self.device, dtype=self.dtype), output_hidden_states=True)
-            image_features = self.feature_select(image_forward_outs).to(images.dtype)
+            # Keep vision tower dtype instead of converting back to input dtype
+            # This is critical for DP training where input dtype might differ from model dtype
+            image_features = self.feature_select(image_forward_outs).to(self.dtype)
 
         return image_features
 
@@ -127,7 +131,8 @@ class CLIPVisionTowerS2(CLIPVisionTower):
     @torch.no_grad()
     def forward_feature(self, images):
         image_forward_outs = self.vision_tower(images.to(device=self.device, dtype=self.dtype), output_hidden_states=True)
-        image_features = self.feature_select(image_forward_outs).to(images.dtype)
+        # Keep vision tower dtype for DP training compatibility
+        image_features = self.feature_select(image_forward_outs).to(self.dtype)
         return image_features
 
     @torch.no_grad()
